fix compiler warning when using OpenMP
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@16814 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: d5fc9803e5b2f33fc848e5a664e5741160363cf7

+++ util/new_robot_self_filter/self_see_filter.h
@@ -163,7 +163,7 @@ public:
 	  std::vector<bool> keep(np);
 
 #pragma omp parallel for 
-	  for (unsigned int i = 0 ; i < np ; ++i)
+	  for (int i = 0 ; i < (int)np ; ++i)
 	  {
 	      btVector3 pt = btVector3(btScalar(data_in.pts[i].x), btScalar(data_in.pts[i].y), btScalar(data_in.pts[i].z));
 	      bool out = true;
